pyproject.toml
[project]
name = "arxiv-agent"
version = "0.1.0"
description = "An AI agent to research latest arxiv papers and send digest to email or webhook"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.0",
    "PyPDF2>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arxiv_agent"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
]

